Making NpgsqlExtendedSqlType serializable - closes #1
