+++ src/nhibernate.postgresql.json.sln
@@ -5,6 +5,8 @@ VisualStudioVersion = 15.0.28010.2003
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "nhibernate.postgresql.json", "nhibernate.postgresql.json\nhibernate.postgresql.json.csproj", "{B260A7D9-805F-4FD3-9A66-E7BF720AB221}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "nhibernate.postgresql.json.tests", "nhibernate.postgresql.json.tests\nhibernate.postgresql.json.tests.csproj", "{0B5E9D55-458C-4B2B-8953-37607AC73ED4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +17,10 @@ Global
 		{B260A7D9-805F-4FD3-9A66-E7BF720AB221}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B260A7D9-805F-4FD3-9A66-E7BF720AB221}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B260A7D9-805F-4FD3-9A66-E7BF720AB221}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0B5E9D55-458C-4B2B-8953-37607AC73ED4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0B5E9D55-458C-4B2B-8953-37607AC73ED4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0B5E9D55-458C-4B2B-8953-37607AC73ED4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0B5E9D55-458C-4B2B-8953-37607AC73ED4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
